move AltNav to its own component
